bType made signed

+++ simulator/src/rTypes.cpp
@@ -33,7 +33,7 @@ void sll(line &instr, uint32_t * reg_ptr){
 	rd = reg_ptr+((instr.instr & EXTRACT_R_RD)>>7);
 	r1 = reg_ptr+((instr.instr & EXTRACT_RBS_R1)>>15);
 	r2 = reg_ptr+((instr.instr & EXTRACT_RBS_R2)>>20);
-	*rd = *r1<<(*r2);
+	*rd = (int32_t)*r1<<((int32_t)*r2);
 	printf("sll x%.2d, x%.2d, x%.2d", ((instr.instr & EXTRACT_R_RD)>>7) , ((instr.instr & EXTRACT_RBS_R1)>>15) , ((instr.instr & EXTRACT_RBS_R2)>>20));
 }
